Support Github Personal Access Token (PAT) based auth for CMS
New `pat_url` config option for Sveltia-CMS. When this config option is enabled, the URL points to a secure endpoint that returns a Git personal access token. When this call is successful, the returned token is injected as if it were a `cachedToken` within the *existing* CMS backend automatic `signIn` flow.  On success, the CMS has the required backend auth token for normal operations exactly as if it had obtained the token from `LocalStorage`. Cached tokens and PAT tokens both share one key characteristic: they both bypass any need for OAuth flow that engages end-users directly. This change intends to deliver an auth capability similar to the now deprecated Netlify Identity service powered by the old `git-gateway` backend. The motivation is to provide a mechanism that will allow Sveltia-CMS deployments give **non-Github users** seamless content authoring access within the CMS.